refactor: separate `velero` module folder
