+++ src/velero/__init__.py
@@ -1,0 +1,12 @@
+"""Velero module."""
+
+from .velero import Velero, VeleroError
+from .velero_providers import AzureStorageProvider, S3StorageProvider, StorageProviderError
+
+__all__ = [
+    "Velero",
+    "VeleroError",
+    "S3StorageProvider",
+    "AzureStorageProvider",
+    "StorageProviderError",
+]
